Rename module subdirectories.
