+++ .github/workflows/check_and_tests.yml
@@ -41,4 +41,4 @@ jobs:
         run: |
           python3 -m pip install nnabla
           python3 -m pip install --only-binary :all: janome nltk scikit-learn plotly matplotlib scikit-image opencv-python
-          bash tools/scripts/check_plugin_help.sh
+          python3 tools/scripts/check_plugin_help.py
