+++ plugins/_Post_Process/_XAI/tracin_utils/args.py
@@ -11,8 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import argparse
-from tracin.utils import ensure_dir
+from .utils import ensure_dir
 from distutils.util import strtobool
