--- tools/scripts/check_plugin_help.sh
@@ -1,25 +1,0 @@
-# Copyright 2021 Sony Group Corporation.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-set -ex
-pip install nnabla
-pip install --only-binary :all: janome nltk scikit-learn plotly matplotlib scikit-image opencv-python
-
-cd $(dirname $0)/../..
-for f in $(find plugins -maxdepth 3 -type f -name "*.py")
-do
-    echo $f
-    python $f --help
-done
-
-
